requirements.txt
aiohttp>=3.9.3
aiosignal>=1.3.1
alabaster>=0.7.16
anyio>=4.2.0
astunparse>=1.6.3
attrs>=23.2.0
Babel>=2.14.0
braceexpand>=0.1.7
Brotli>=1.1.0
certifi>=2023.11.17
cffi>=1.16.0
characterai>=0.8.0
chardet>=5.2.0
charset-normalizer>=3.3.2
click>=8.1.7
discord.py>=2.3.2
dnspython>=2.5.0
docutils>=0.20.1
frozenlist>=1.4.1
gitdb>=4.0.11
GitPython>=3.1.41
humanize>=4.9.0
idna>=3.6
imagesize>=1.4.1
import-expression>=1.1.4
Jinja2>=3.1.3
jishaku>=2.5.2
markdown-it-py>=3.0.0
MarkupSafe>=2.1.4
mdurl>=0.1.2
motor>=3.3.2
multidict>=6.0.5
mutagen>=1.47.0
mystbin-py>=6.1.0
packaging>=23.2
pillow>=10.2.0
psutil>=5.9.8
pycparser>=2.21
pycryptodomex>=3.20.0
Pygments>=2.17.2
pymongo>=4.6.1
PyNaCl>=1.5.0
python-dateutil>=2.8.2
requests>=2.31.0
rich>=13.7.0
setuptools>=69.0.3
six>=1.16.0
smmap>=5.0.1
sniffio>=1.3.0
snowballstemmer>=2.2.0
Sphinx>=7.2.6
sphinx-rtd-dark-mode>=1.3.0
sphinx-rtd-theme>=2.0.0
sphinxcontrib-applehelp>=1.0.8
sphinxcontrib-devhelp>=1.0.6
sphinxcontrib-htmlhelp>=2.0.5
sphinxcontrib-jquery>=4.1
sphinxcontrib-jsmath>=1.0.1
sphinxcontrib-qthelp>=1.0.7
sphinxcontrib-serializinghtml>=1.1.10
tls-client>=1.0
typing_extensions>=4.9.0
urllib3>=2.1.0
waifuim.py>=4.2.4
websockets>=12.0
wheel>=0.42.0
yarl>=1.9.4
